+++ src/CoreBackend/Features/submissions/SubmissionsController.cs
@@ -17,14 +17,19 @@ namespace CoreBackend.Features.Submissions
             _submissionsService = submissionsService;
         }
 
+        [HttpGet]
+        public async Task<ActionResult<List<GetSubmissionRO>>> GetSeveralSubmissions([FromQuery] SubmissionsFilterDTO submissionsFilter)
+        {
+            return await _submissionsService.GetSeveralSubmissions(submissionsFilter);
+        }
+
         [HttpPost]
-        // [RequireRole(UserRole.Student)]
+        [RequireRole(UserRole.Student)]
         public async Task<ActionResult<CreateSubmissionRO>> CreateSubmission([FromBody] CreateSubmissionDTO createSubmissionDTO)
         {
-            // UserRole? role = HttpContext.Items[Consts.HTTP_CONTEXT_USER_ROLE] as UserRole?;
-            // if (role == null)
-            //     return Forbid();
-            Console.WriteLine(createSubmissionDTO.personalNum);
+            UserRole? role = HttpContext.Items[Consts.HTTP_CONTEXT_USER_ROLE] as UserRole?;
+            if (role == null)
+                return Forbid();
             return await _submissionsService.CreateSubmission(createSubmissionDTO);
         }
 
@@ -35,14 +40,23 @@ namespace CoreBackend.Features.Submissions
         }
 
         [HttpPatch("{submissionID}")]
+        [Consumes("application/json")]
+        [RequireRole(UserRole.Admin, UserRole.Commander, UserRole.Mamak, UserRole.Student)]
         public async Task<ActionResult> UpdateSpecificSubmission(string submissionID, [FromBody] JsonElement updateElement)
         {
+            UserRole? role = HttpContext.Items[Consts.HTTP_CONTEXT_USER_ROLE] as UserRole?;
+            if (role == null)
+                return Forbid();
             return await _submissionsService.UpdateSpecificSubmission(submissionID, updateElement);
         }
 
         [HttpDelete("{submissionID}")]
+        [RequireRole(UserRole.Admin, UserRole.Commander, UserRole.Mamak, UserRole.Student)]
         public async Task<ActionResult> RemoveSpecificSubmission(string submissionID)
         {
+            UserRole? role = HttpContext.Items[Consts.HTTP_CONTEXT_USER_ROLE] as UserRole?;
+            if (role == null)
+                return Forbid();
             return await _submissionsService.RemoveSpecificSubmission(submissionID);
         }
 
@@ -51,6 +65,5 @@ namespace CoreBackend.Features.Submissions
         {
             return await _submissionsService.IsSubmissionApproved(submissionID);
         }
-
     }
 }
